requirements.txt
autopep8
fake-bpy-module-latest